requirements.txt
python-telegram-bot>=20.0
PyQt5>=5.15.0
playwright>=1.40.0
requests>=2.30.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
spacy>=3.7.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
rich>=13.0.0
aiohttp>=3.8.0
Flask>=2.3.0
